Fix compilation/rendering when extensions are involved
In Scala3, in the case of extension methods, the sub-type-tree that the
macro is trying to capture to pass to the recorder doesn't quite match
what the recorder expects, leading to a compile error.

In addition (in all scala versions), extension class instances are
captured by the macro leading to useless/irrelevant rendering.

The changes here attempt to detect implicit applications when traversing
the tree to skip them during recording.